// api/interfaces/userInterface.ts
import { genderEnum } from "../enums/genderEnum";
import { Roles } from "../enums/roleEnum";

export interface UserInterface {
    id: string;
    idCard: number;
    firstName: string;
    lastName: string;
    gender: genderEnum | undefined;
    country: string;
    state: string;
    city: string;
    address: string;
    postalCode: number;
    email: string;
    phone: number;
    dateBirth: string | Date | undefined;
    createdAt: Date;
    updatedAt: Date;
    role: Roles | undefined;
    password: string;

}